feat: add reusable SequenceList component
